+++ core/src/main/java/com/epam/esm/service/impl/UserServiceImpl.java
@@ -25,15 +25,15 @@ public class UserServiceImpl implements UserService {
     public CustomPage<UserDto> getPaginated(CustomPageable pageRequest) {
         int size = pageRequest.getSize();
         int page = pageRequest.getPage();
-        int totalUserAmount = userRepository.countAll();
-        int lastPage = (totalUserAmount + size - 1) / size - 1;
+        long totalUserAmount = userRepository.countAll();
+        long lastPage = (totalUserAmount + size - 1) / size - 1;
         if (page > lastPage) {
             throw new ResourceNotFoundException(String.format(ErrorMessage.PAGE_NOT_FOUND, size, page, lastPage));
         }
         int offset = size * page;
         List<User> foundUsers = userRepository.findAllPaginated(offset, size);
         return new CustomPage<>(foundUsers.stream().map(userConverter::toDTO).collect(Collectors.toList()),
-                pageRequest, (long) totalUserAmount);
+                pageRequest, totalUserAmount);
     }
 
     @Override
